+++ src/main/resources/sql/recommend.sql
@@ -31,9 +31,9 @@ CREATE TABLE laptop_data(
 --감정 분석 테이블
 CREATE TABLE laptop_sentiment (
     product_no INT NOT NULL ,
-    sentiment_positive DOUBLE NOT NULL,
-    sentiment_denial DOUBLE NOT NULL,
-    sentiment_neutrality DOUBLE NOT NULL,
+        sentiment_positive INT NOT NULL,
+    sentiment_denial INT NOT NULL,
+    sentiment_neutrality INT NOT NULL,
     primary key (product_no),
     FOREIGN KEY (product_no) REFERENCES product(product_no)
 );
